+++ examples/async_api_patterns.rs
@@ -7,7 +7,7 @@
 //!
 //! Run with: `cargo run --example async_api_patterns --features async-full,ecosystem`
 
-#![cfg(feature = "ecosystem")]
+#![cfg(feature = "tokio")]
 
 use error_rail::context;
 use error_rail::prelude_async::*;