+++ tests/tower_tests.rs
@@ -1,5 +1,5 @@
 //! Tests for Tower integration.
-#![cfg(feature = "ecosystem")]
+#![cfg(feature = "tower")]
 
 use std::future::Future;
 use std::pin::Pin;